refactor: move mocks to a dedicated mocks directory
