+++ tests/mocks/user-icon-collection.service.mock.ts
